feat: Support logging config.
The logging config supports configuration of the logging destination,
and the maximum logging level options currently.

BREAKING CHANGE:

Note that NewLogger() function signature changes to allow supplying the
logging config. Any additional configuration options in the future will
not be breaking and will be extensions to the same Config struct.